Added separate actions for tests & code coverage, PMD and checkstyle.
Closes #5.
Retained the end to end builds and added separate actions for tests & code coverage, PMD and checkstyle.
This is done so that it is easier to figure out the check that caused the failure in the end to end build as well
as to collect metrics on the checks after some time.